근무 그리드 컬럼 수정, 중복근무 신청 기각, 팝업 새로고침 이슈 해결

+++ src/main/java/com/itwillbs/factron/repository/test/TestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestRepository extends JpaRepository<Employee, Integer> {
+public interface TestRepository extends JpaRepository<Employee, Long> {
     // 삽입, 수정 시 사용
 }
